- mtca4u.read noe returns array type numpy.float64(double); Previous behaviour   was to return the read value in a numpy.float32 array.
git-svn-id: https://svnsrv.desy.de/desy/mtca4u/PythonBindings/deviceaccess/trunk@2246 5f601a5c-88b9-47f3-b851-818f6741d00d

+++ RegisterAccessor.cc
@@ -8,10 +8,10 @@ namespace RegisterAccessor {
             bp::numeric::array& numpyArray, size_t numberOfElementsToRead,
             uint32_t startIndexInRegister) {
 
-    if (extractDataType(numpyArray) == FLOAT32) {
-      float* allocatedSpace =
-          reinterpret_cast<float*>(extractDataPointer(numpyArray));
-      self.read<float>(allocatedSpace, numberOfElementsToRead,
+    if (extractDataType(numpyArray) == FLOAT64) {
+      double* allocatedSpace =
+          reinterpret_cast<double*>(extractDataPointer(numpyArray));
+      self.read<double>(allocatedSpace, numberOfElementsToRead,
                        startIndexInRegister);
     } else {
       throw mtca4upy::ArrayElementTypeNotSupported(
